fix: refactor AuthController to handle login in AuthService

+++ src/main/java/com/example/formula1_demo/service/DriverService.java
@@ -3,6 +3,7 @@ package com.example.formula1_demo.service;
 import com.example.formula1_demo.entity.Driver;
 import com.example.formula1_demo.repository.DriverRepository;
 
+import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -10,56 +11,69 @@ import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
-public class DriverService {
+@AllArgsConstructor
+public class DriverService implements IDriverService {
 
-    private DriverRepository driverRepository;
+    private final DriverRepository driverRepository;
 
-    public DriverService(DriverRepository driverRepository) {
-        this.driverRepository = driverRepository;
-    }
-
+    @Override
     public List<Driver> getAllDrivers() {
         return driverRepository.findAll();
     }
 
+    @Override
     public Optional<Driver> getDriverById(Long id) {
         return driverRepository.findById(id);
     }
 
+    @Override
     public Optional<List<Driver>> getDriverByName(String name) {
         List<Driver> drivers = driverRepository.findAll().stream()
-                .filter(driver -> driver.getDriver().toLowerCase().contains(name.toLowerCase())).collect(Collectors.toList());
+                .filter(driver -> driver.getDriver().toLowerCase().contains(name.toLowerCase()))
+                .collect(Collectors.toList());
         return Optional.of(drivers);
     }
 
+    @Override
     public Optional<List<Driver>> getDriverByNationality(String nationality) {
         List<Driver> drivers = driverRepository.findAll().stream()
-                .filter(driver -> driver.getNationality().toLowerCase().equals(nationality.toLowerCase())).collect(Collectors.toList());
+                .filter(driver -> driver.getNationality().equalsIgnoreCase(nationality))
+                .collect(Collectors.toList());
         return Optional.of(drivers);                
     }
 
+    @Override
     public Optional<List<Driver>> getDriversByNameAndNationality(String name, String nationality) {
         List<Driver> drivers = driverRepository.findAll().stream()
-                .filter(driver -> driver.getDriver().toLowerCase().contains(name.toLowerCase()) && driver.getNationality().toLowerCase().contains(nationality.toLowerCase())).collect(Collectors.toList());
+                .filter(driver -> driver.getDriver().toLowerCase().contains(name.toLowerCase()) &&
+                                  driver.getNationality().equalsIgnoreCase(nationality))
+                .collect(Collectors.toList());
         return Optional.of(drivers);
     }
 
+    @Override
     public Optional<List<Driver>> getDriverBySeasonsActive(String seasons) {
         List<Driver> drivers = driverRepository.findAll().stream()
-                .filter(driver -> driver.getSeasons().toLowerCase().contains(seasons.toLowerCase())).collect(Collectors.toList());
+                .filter(driver -> driver.getSeasons().toLowerCase().contains(seasons.toLowerCase()))
+                .collect(Collectors.toList());
         return Optional.of(drivers);
     }
 
+    @Override
     public Optional<List<Driver>> getDriverByChampion(Boolean champion) {
         List<Driver> drivers = driverRepository.findAll().stream()
-                .filter(driver -> driver.getChampion().equals(champion)).collect(Collectors.toList());
+                .filter(driver -> driver.getChampion().equals(champion))
+                .collect(Collectors.toList());
         return Optional.of(drivers);
     }
 
+    @Override
     public Driver createDriver(Driver driver) {
+        
         return driverRepository.save(driver);
     }
 
+    @Override
     public Driver updateDriver(Long id, Driver driver) {
         return driverRepository.findById(id)
                 .map(existingDriver -> {
@@ -93,7 +107,14 @@ public class DriverService {
                 });
     }
 
+    @Override
     public void deleteDriver(Long id) {
-        driverRepository.deleteById(id);
+        Optional<Driver> driver = driverRepository.findById(id);
+        if (driver.isPresent()) {
+            driverRepository.deleteById(id);
+        } else {
+            throw new RuntimeException("Driver with ID " + id + " not found.");
+        }
     }
+
 }
